+++ lib/utils/app_variables.dart
@@ -8,9 +8,9 @@ class AppVariables {
   static const double SLIDER_HEIGHT = 44;
   static const double MAIN_BORDER_RADIUS = 4;
 
-  static const double MAX_TIME = 60;
-  static const double INIT_TIME = 30;
-  static const double EXTEND_TIME = 5;
+  static const int MAX_TIME = 60;
+  static const int INIT_TIME = 30;
+  static const int EXTEND_TIME = 5;
 
   static final navigatorKey = GlobalKey<NavigatorState>();
 }
